refactor code to reuse the method that create a order in fulfillment

+++ src/main/java/com/inretailpharma/digital/deliverymanager/entity/projection/IOrderFulfillment.java
@@ -10,7 +10,7 @@ public interface IOrderFulfillment {
     Long getEcommerceId();
     Long getTrackerId();
     Long getExternalId();
-    Integer getPurchaseId();
+    Long getPurchaseId();
     String getExternalChannelId();
